requirements.txt
# Core AI and ML
langchain==0.1.0
langchain-openai==0.0.5
langchain-google-genai==0.0.6
langgraph==0.0.20
openai==1.12.0
google-generativeai==0.3.2

# FastAPI and Django
fastapi==0.116.1
uvicorn[standard]==0.35.0
django==5.2.4
djangorestframework==3.15.0
django-cors-headers==4.3.1

# Database
asyncpg==0.30.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.25
alembic==1.13.1

# Financial APIs
yfinance==0.2.28
alpha-vantage==2.3.1
requests==2.31.0

# Firebase
firebase-admin==6.4.0

# Data processing
pandas==2.3.2
numpy==1.26.2
python-dotenv==1.0.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
pydantic==2.11.7
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1 